requirements.txt
flask
curl_cffi
beautifulsoup4
python-dotenv
azure-ai-projects
azure-identity